feat: provider 함수 수정
파일명, 함수명, 폴더구조 변경

Related to: #89
