+++ lib/provider/NextUIProvider.tsx
@@ -1,0 +1,10 @@
+'use client';
+
+import { NextUIProvider as Provider } from '@nextui-org/react';
+import { useRouter } from 'next/navigation';
+
+export default function NextUIProvider({ children }: { children: React.ReactNode }) {
+  const router = useRouter();
+
+  return <Provider navigate={router.push}>{children}</Provider>;
+}
